Changed type of iterations to have them more.

+++ hill_climbing.cpp
@@ -67,7 +67,7 @@ namespace marxan {
 
             }
 
-            void append(int i, int puno, const scost& reserve, const scost& change, const vector<int>& R)
+            void append(long long i, int puno, const scost& reserve, const scost& change, const vector<int>& R)
             {
                 iRowCounter++;
                 if (iRowCounter > iRowLimit)
@@ -75,9 +75,9 @@ namespace marxan {
 
                 if (iRowCounter == 1)
                 {
-                    fprintf(Rfp, "%i", i);
+                    fprintf(Rfp, "%lli", i);
 
-                    fprintf(ttfp, "%i,%f,%i,%f,%f,%f,%f\n"
+                    fprintf(ttfp, "%lli,%f,%i,%f,%f,%f,%f\n"
                                 , i, reserve.total
                                 , reserve.pus, reserve.cost, reserve.connection, reserve.penalty,
                                 change.total); // i,costthresh,pus,cost,connection,penalty
@@ -121,7 +121,7 @@ namespace marxan {
     void hill_climbing(int puno, int spno, const vector<spustuff>& pu, const vector<sconnections>& connections,
         vector<sspecies>& spec, const vector<spu>& SM, vector<spu_out>& SM_out, vector<int>& R, double cm,
         scost& reserve, double costthresh, double tpf1, double tpf2,
-        int clumptype,  int irun, int iterations, string savename, stringstream& logBuffer, rng_engine& rngEngine)
+        int clumptype,  int irun, long long iterations, string savename, stringstream& logBuffer, rng_engine& rngEngine)
     {
         scost change = { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 };
         int puvalid = 0,  ipu = 0, imode, ichoice;
@@ -156,7 +156,7 @@ namespace marxan {
             logBuffer << "Hill climbing after array init\n";
             displayProgress2("  Main Hillclimbing Section.\n");
 
-            for(int itime = 1; itime <= iterations; )
+            for(long long itime = 1; itime <= iterations; )
             {
                 // shuffle iimp array
                 std::shuffle(iimparray.begin(), iimparray.end(), rngEngine);
@@ -196,7 +196,7 @@ namespace marxan {
     void hill_climbing_two_steps(int puno, int spno, const vector<spustuff>& pu, const vector<sconnections>& connections,
         vector<sspecies>& spec, const vector<spu>& SM, vector<spu_out>& SM_out, vector<int>& R, double cm,
         scost& reserve, double costthresh, double tpf1, double tpf2,
-        int clumptype,  int irun, int iterations, string savename, stringstream& logBuffer, rng_engine& rngEngine)
+        int clumptype,  int irun, long long iterations, string savename, stringstream& logBuffer, rng_engine& rngEngine)
     {
         int puvalid = 0,  ipu = 0;
         vector<int> iimparray;
@@ -231,7 +231,7 @@ namespace marxan {
             displayProgress2("  Main two step hillclimbing section.\n");
 
              
-            for(int itime = 1; itime <= iterations; )
+            for(long long itime = 1; itime <= iterations; )
             {
                 // shuffle iimp array
                 std::shuffle(iimparray.begin(), iimparray.end(), rngEngine);
